requirements.txt
boto3==1.14.12
click==7.1.2
cycler==0.10.0
docutils==0.15.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.2
mysql-connector-python==8.0.20
numpy==1.19.0
opencv-python==4.2.0.34
pandas==1.0.5
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sklearn==0.0
urllib3==1.25.9
Werkzeug==1.0.1
yattag==1.13.2
botocore==1.17.13
protobuf==3.12.2
s3transfer==0.3.3